Remove faulty asserts
We were fundamentally wrong about what a 16 cell is. Previously, it was
thought that a 16 cell is a cell whose north west corner is an 8 cell
(i.e. a leaf) but this is not quite correct. Since a void cell is not
considered a leaf, a 16 cell with a void north west corner would not
have been considered a 16 cell by the previous code's logic. The logic
was updated to define a 16 cell as a cell with *at least* one leaf
quadrant (The implication being that the others could only be voids).